# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 808 All Purpose Cleaners Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /checklist/{id}:
    patch:
      operationId: updateChecklistItem
      tags: [checklist]
      summary: Toggle or update a checklist item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChecklistItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistItem"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, createdAt, updatedAt]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
      required: [name]
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Booking:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        serviceType:
          type: string
        status:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        estimatedPrice:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, serviceType, status, date, time, createdAt, updatedAt]
    CreateBookingBody:
      type: object
      properties:
        clientId:
          type: integer
        employeeId:
          type: integer
        serviceType:
          type: string
        status:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        notes:
          type: string
        estimatedPrice:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
      required: [serviceType, date, time]
    UpdateBookingBody:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        serviceType:
          type: string
        status:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        estimatedPrice:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, status, createdAt, updatedAt]
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: string
        targetAudience:
          type: string
        description:
          type: string
        notes:
          type: string
      required: [name, type]
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        completed:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, category, title, completed, sortOrder, createdAt, updatedAt]
    UpdateChecklistItemBody:
      type: object
      properties:
        completed:
          type: boolean
...