# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 808 All Purpose Cleaners Business Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  employeeId?: number | null;
  serviceType: string;
  status: string;
  date: string;
  time: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  estimatedPrice?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientPhone?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  createdAt: Date;
  updatedAt: Date;
}